Move fixtures to top-level of tests module
